+++ src/main/java/com/subu/fileprocessor/dao/Offset.java
@@ -3,13 +3,11 @@ package com.subu.fileprocessor.dao;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
-import java.math.BigInteger;
-
 
 @Data
 @RequiredArgsConstructor
 public class Offset {
-    private final BigInteger lineOffset;
-    private final BigInteger charOffset;
+    private final long lineOffset;
+    private final long charOffset;
 }
